Internal: Update paragraph [ED-16740]
